+++ CHANGELOG.md
@@ -3,5 +3,14 @@
 All notable changes to this project are documented in this file.
 
 ## [1.0.0] - 26-11-2025
-
 - initialized the repository
+
+## [1.1.0] - 03-12-2025
+### Added
+- **Flask** based web dashboard
+- Added the following modules:
+  - Photo Resistor
+  - PIR sensor
+  - Relay module
+  - DH11
+- Added LCD display (16x2 I2C added)